Refactoring raw atlas slicer into two distinct pieces - one for slicing relations, the other for lines/points, updating Iterables and test
